Handle promise bun run exec
