+++ packages/models/src/bun/index.ts
@@ -1,0 +1,10 @@
+export interface RunBunCommandOptions {
+    cwd: string;
+    env?: NodeJS.ProcessEnv;
+}
+
+export interface RunBunCommandResult {
+    success: boolean;
+    error?: string;
+    output?: string;
+}
