+++ packages/models/src/index.ts
@@ -1,4 +1,5 @@
 export * from './actions/';
+export * from './bun/';
 export * from './chat/';
 export * from './code/';
 export * from './constants/';
@@ -6,6 +7,7 @@ export * from './create/';
 export * from './editor/';
 export * from './element/';
 export * from './ide/';
-export * from './projects/';
-export * from './settings/';
 export * from './pages/';
+export * from './projects/';
+export * from './run/';
+export * from './settings/';
